consolidate accessor GetApplication methods.
request handlers now convert from the accessor response to the client response, rather than the accessor trying to build the correct client response (if that makes sense).

the notes are now directly on the application details contract, but are never serialized and thus are never returned to the client. this is perfect for the applicant view, but for the reviewer view the notes are transferred from the application details to the response contract itself.